GUI/requirements.txt
PySide6
PyOpenGL
PyWavefront
numpy
pymongo
dnspython
pyassimp
pyqtgraph
python-dotenv
bleak
pandas
tensorflow
scikit-learn
python-dateutil
joblib